Extract reusable PipeTransfer utilities
Move high-level pipe copy and wait expansion out of the TTL conversion pass so earlier PipeNet transformations can consume the same explicit transfer IR.

Centralize checked sender-payload geometry in PipePlanning. Later transport grouping can extend the number of consecutive DFB blocks without replacing the scalar payload calculation.

Keep selected-record support and transfer provenance handling in the shared expansion implementation.
