+++ lib/Dialect/TTL/Transforms/CMakeLists.txt
@@ -17,6 +17,7 @@ add_mlir_dialect_library(TTLangTTLTransforms
   PipeLowering.cpp
   PipePlanning.cpp
   PipeTransferAnalysis.cpp
+  PipeTransferExpansion.cpp
   ConvertTTLToCompute.cpp
   ConvertTTLComputeToSCF.cpp
   ConvertTTLTileOpsToTTKernel.cpp
